Fix (MSVC) compiler warnings
reg. size_t vs. int, after #11194

+++ src/engraving/libmscore/harmony.cpp
@@ -2008,7 +2008,7 @@ void Harmony::addDegree(const HDegree& d)
 //   numberOfDegrees
 //---------------------------------------------------------
 
-int Harmony::numberOfDegrees() const
+size_t Harmony::numberOfDegrees() const
 {
     return _degreeList.size();
 }
